edge/pyproject.toml
[project]
name = "elderos-edge"
version = "0.1.0"
description = "ElderOS Vision Edge Appliance Software"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11",
    "opencv-python-headless>=4.10",
    "onnxruntime>=1.20",
    "pyyaml>=6.0",
    "ntplib>=0.4",
    "aiosqlite>=0.20",
    "pydantic>=2.10",
    "psutil>=6.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]
serial = ["pyserial-asyncio>=0.6"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
